Tests/LibWeb/Text/input/css/subgrid-auto-fill-reserves-trailing-line-names.html

<!DOCTYPE html>
<style>
    body {
        margin: 0;
    }

    .grid {
        display: grid;
        grid-template-columns: repeat(4, 20px);
    }

    .subgrid {
        display: grid;
        grid-column: 1 / span 4;
        grid-template-columns: subgrid [a] repeat(auto-fill, [b]) [c];
    }
</style>
<div class="grid"><div class="subgrid"></div></div>
<script src="../include.js"></script>
<script>
    test(() => {
        println(getComputedStyle(document.querySelector(".subgrid")).gridTemplateColumns);
    });
</script>
