Tests/LibWeb/Text/expected/css/subgrid-auto-fill-reserves-trailing-line-names.txt

subgrid [a] [b] [b] [b] [c]
